+++ dask_geomodeling/tests/test_raster_spatial.py
@@ -205,3 +205,43 @@ def test_place_time_request(source, center):
 def test_place_meta_request(source, center):
     place = raster.Place(source, "EPSG:28992", center, [(150, 50)])
     assert source.get_data(mode="meta") == place.get_data(mode="meta")
+
+
+@pytest.mark.parametrize("point,expected", [
+    ((5, 15), 7),  # box 1
+    ((15, 15), 255),  # box 2
+    ((5, 5), 255),  # box 3
+    ((15, 5), 7),  # box 4
+    ((10, 15), 255),  # line 1-2
+    ((5, 10), 255),  # line 1-3
+    ((15, 10), 7),  # line 2-4
+    ((10, 5), 7),  # line 3-4
+    ((10, 10), 7),  # center
+])
+def test_place_point_request(source, center, point, expected):
+    # For point requests, edges are important. Let's do a drawing:
+    # 20  _______ _______
+    #    |       |       |
+    #    |   1   |   2   |
+    # 10 |_______|_______|
+    #    |       |       |
+    #    |   3   |   4   |
+    #  0 |_______|_______|
+    #    0       10     20
+    # - boxes 1 and 4 are filled; box 2 and 3 are empty (see below coordinates)
+    # A cell includes its topleft corner and top and left edges
+    # - line between 2-4 and 3-4 are filled; line 1-2 and 1-3 are empty
+    # - center point at (10, 10) is filled
+    coordinates = [(60, -40), (-40, 60)]
+    place = raster.Place(source, "EPSG:28992", center, coordinates)
+    point_request = dict(
+        mode="vals",
+        bbox=point * 2,
+        projection="EPSG:28992",
+        width=1,
+        height=1,
+    )
+    values = place.get_data(**point_request)["values"]
+    # swap Y axis for readable test
+    assert values.shape == (1, 1, 1)
+    assert values.item() == expected